cuda - separate compile functionality into new header
