+++ backends/cuda-gen/ceed-cuda-gen-qfunction.c
@@ -20,7 +20,6 @@
 #include <stdio.h>
 #include <string.h>
 #include "ceed-cuda-gen.h"
-#include "../cuda/ceed-cuda.h"
 
 //------------------------------------------------------------------------------
 // Apply QFunction
