+++ backends/cuda/ceed-cuda-qfunction.c
@@ -20,6 +20,7 @@
 #include <stdio.h>
 #include <string.h>
 #include "ceed-cuda.h"
+#include "ceed-cuda-compile.h"
 #include "ceed-cuda-qfunction-load.h"
 
 //------------------------------------------------------------------------------
